PWN-7039 - deprecated old, fix constrain, moved icon wrapper
