+++ ui-kit/src/main/kotlin/org/p2p/uikit/atoms/UiKitEndAmountView.kt
@@ -12,6 +12,7 @@ import org.p2p.uikit.utils.inflateViewBinding
 import org.p2p.uikit.utils.withImageOrGone
 import org.p2p.uikit.utils.withTextOrGone
 
+@Deprecated("use [UiKitFinanceBlockView] without [FinanceBlockUiModel.leftSideUiModel]")
 class UiKitEndAmountView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
