+++ ui-kit/src/main/kotlin/org/p2p/uikit/components/finance_block/UiKitFinanceBlockView.kt
@@ -1,7 +1,6 @@
 package org.p2p.uikit.components.finance_block
 
 import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.core.view.isInvisible
 import androidx.core.view.isVisible
 import androidx.core.view.updatePadding
 import android.content.Context
@@ -40,7 +39,7 @@ class UiKitFinanceBlockView @JvmOverloads constructor(
         _item = model
         isEnabled = model.accessibility.isEnabled
         isClickable = model.accessibility.isClickable
-        binding.leftSideView.isInvisible = model.leftSideUiModel == null
+        binding.leftSideView.isVisible = model.leftSideUiModel != null
         model.leftSideUiModel?.let { binding.leftSideView.bind(it) }
         binding.rightSideView.isVisible = model.rightSideUiModel != null
         model.rightSideUiModel?.let { binding.rightSideView.bind(it) }
