+++ ui-kit/src/main/kotlin/org/p2p/uikit/components/left_side/LeftSideUiModel.kt
@@ -1,6 +1,6 @@
 package org.p2p.uikit.components.left_side
 
-import org.p2p.uikit.atoms.icon_wrapper.IconWrapperUiModel
+import org.p2p.uikit.components.icon_wrapper.IconWrapperUiModel
 import org.p2p.uikit.utils.text.TextViewUiModel
 
 sealed interface LeftSideUiModel {
