+++ ui-kit/src/main/res/layout/widget_icon_wrapper_single.xml
@@ -13,6 +13,7 @@
         android:id="@+id/imageViewIcon"
         android:layout_width="0dp"
         android:layout_height="0dp"
+        android:duplicateParentState="true"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
