+++ ui-kit/src/main/res/layout/widget_icon_wrapper_two.xml
@@ -13,6 +13,7 @@
         android:id="@+id/imageViewFirstIcon"
         android:layout_width="0dp"
         android:layout_height="0dp"
+        android:duplicateParentState="true"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHeight_percent="0.625"
@@ -27,6 +28,7 @@
         android:id="@+id/imageViewSecondIcon"
         android:layout_width="0dp"
         android:layout_height="0dp"
+        android:duplicateParentState="true"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHeight_percent="0.625"
