+++ ui-kit/src/main/res/layout/widget_right_side_double_text.xml
@@ -13,6 +13,7 @@
         android:id="@+id/textViewFirst"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:duplicateParentState="true"
         android:ellipsize="end"
         android:maxLines="1"
         android:textAppearance="@style/UiKit.TextAppearance.SemiBold.Text3"
@@ -31,6 +32,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="4dp"
+        android:duplicateParentState="true"
         android:ellipsize="end"
         android:maxLines="1"
         android:textAppearance="@style/UiKit.TextAppearance.Regular.Label1"
